Explicitly sum counts in order to arrange bars
Tbh I think summing the counts across years is not a great way to look at this data, as it obscures the fact that the number of people in the sample is different each year.  But he version we're publishing looks at each year seaprately, which is a better move.